inline test class into test project
